router: routes and functionality separated
