+++ src/index.js
@@ -1,4 +1,4 @@
-import router from './routes/index.js';
+import router from './router/index.js';
 
 window.addEventListener('load', router);
 window.addEventListener('hashchange', router);
