+++ src/router/routes.js
@@ -1,0 +1,11 @@
+import Home from '../pages/Home.js';
+import Blog from '../pages/Blog.js';
+import Collaborator from '../pages/Collaborator.js';
+
+const routes = {
+	'/': Home,
+	'/blog': Blog,
+	'/collaborator': Collaborator,
+};
+
+export default routes;
